feature(DEV-889): do not delete tag

+++ providers/aws/vpc/subnet.go
@@ -10,9 +10,10 @@ import (
 )
 
 type Subnet struct {
-	Id string
+	Id           string
 	CreationDate time.Time
-	ttl int64
+	ttl          int64
+	IsProtected  bool
 }
 
 func getSubnetsByVpcId (ec2Session ec2.EC2, vpcId string) []*ec2.Subnet {
@@ -51,19 +52,20 @@ func getSubnetsByVpcsIds (ec2Session ec2.EC2, vpcsIds []*string) []*ec2.Subnet {
 	return result.Subnets
 }
 
-func SetSubnetsIdsByVpcId (ec2Session ec2.EC2, vpc *VpcInfo, waitGroup *sync.WaitGroup) {
+func SetSubnetsIdsByVpcId (ec2Session ec2.EC2, vpc *VpcInfo, waitGroup *sync.WaitGroup, tagName string) {
 	defer waitGroup.Done()
 	var subnetsStruct []Subnet
 
 	subnets := getSubnetsByVpcId(ec2Session, *vpc.VpcId)
 
 	for _, subnet := range subnets {
-		creationDate, ttl := utils.GetTimeInfos(subnet.Tags)
+		creationDate, ttl, isProtected, _, _ := utils.GetEssentialTags(subnet.Tags, tagName)
 
 		var subnetStruct = Subnet{
 			Id: *subnet.SubnetId,
 			CreationDate: creationDate,
 			ttl: ttl,
+			IsProtected: isProtected,
 		}
 		subnetsStruct = append(subnetsStruct, subnetStruct)
 	}
@@ -73,7 +75,7 @@ func SetSubnetsIdsByVpcId (ec2Session ec2.EC2, vpc *VpcInfo, waitGroup *sync.Wai
 
 func DeleteSubnetsByIds (ec2Session ec2.EC2, subnets []Subnet) {
 	for _, subnet := range subnets {
-		if utils.CheckIfExpired(subnet.CreationDate, subnet.ttl) {
+		if utils.CheckIfExpired(subnet.CreationDate, subnet.ttl) && subnet.IsProtected {
 			_, err := ec2Session.DeleteSubnet(
 				&ec2.DeleteSubnetInput{
 					SubnetId: aws.String(subnet.Id),
